chore: refactor unit test struct
