--- pkg/streamx/provider/ttstream/ttstream_common_test.go
@@ -1,61 +1,0 @@
-/*
- * Copyright 2024 CloudWeGo Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ttstream_test
-
-import (
-	"context"
-
-	"github.com/bytedance/gopkg/cloud/metainfo"
-)
-
-var persistKVs = map[string]string{
-	"p1": "v1",
-	"p2": "v2",
-	"p3": "v3",
-}
-
-var transitKVs = map[string]string{
-	"t1": "v1",
-	"t2": "v2",
-	"t3": "v3",
-}
-
-func SetMetadata(ctx context.Context) context.Context {
-	for k, v := range persistKVs {
-		ctx = metainfo.WithPersistentValue(ctx, k, v)
-	}
-	for k, v := range transitKVs {
-		ctx = metainfo.WithValue(ctx, k, v)
-	}
-	return ctx
-}
-
-func ValidateMetadata(ctx context.Context) bool {
-	for k, v := range persistKVs {
-		_v, _ := metainfo.GetPersistentValue(ctx, k)
-		if _v != v {
-			return false
-		}
-	}
-	for k, v := range transitKVs {
-		_v, _ := metainfo.GetValue(ctx, k)
-		if _v != v {
-			return false
-		}
-	}
-	return true
-}
