+++ pkg/streamx/streamx_common_test.go
@@ -1,0 +1,94 @@
+/*
+ * Copyright 2024 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package streamx_test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/bytedance/gopkg/cloud/metainfo"
+	"github.com/cloudwego/gopkg/protocol/thrift"
+	"github.com/cloudwego/kitex/internal/test"
+	"github.com/cloudwego/kitex/pkg/kerrors"
+	"github.com/cloudwego/kitex/pkg/remote"
+)
+
+const (
+	normalErr int32 = iota + 1
+	bizErr
+)
+
+var (
+	testCode  = int32(10001)
+	testMsg   = "biz testMsg"
+	testExtra = map[string]string{
+		"testKey": "testVal",
+	}
+	normalErrMsg = "normal error"
+
+	persistKVs = map[string]string{
+		"p1": "v1",
+		"p2": "v2",
+		"p3": "v3",
+	}
+	transitKVs = map[string]string{
+		"t1": "v1",
+		"t2": "v2",
+		"t3": "v3",
+	}
+)
+
+func setMetadata(ctx context.Context) context.Context {
+	for k, v := range persistKVs {
+		ctx = metainfo.WithPersistentValue(ctx, k, v)
+	}
+	for k, v := range transitKVs {
+		ctx = metainfo.WithValue(ctx, k, v)
+	}
+	return ctx
+}
+
+func validateMetadata(ctx context.Context) bool {
+	for k, v := range persistKVs {
+		_v, _ := metainfo.GetPersistentValue(ctx, k)
+		if _v != v {
+			return false
+		}
+	}
+	for k, v := range transitKVs {
+		_v, _ := metainfo.GetValue(ctx, k)
+		if _v != v {
+			return false
+		}
+	}
+	return true
+}
+
+func assertNormalErr(t *testing.T, err error) {
+	ex, ok := err.(*thrift.ApplicationException)
+	test.Assert(t, ok, err)
+	test.Assert(t, ex.TypeID() == remote.InternalError, ex.TypeID())
+	test.Assert(t, ex.Msg() == "biz error: "+normalErrMsg, ex.Msg())
+}
+
+func assertBizErr(t *testing.T, err error) {
+	bizIntf, ok := kerrors.FromBizStatusError(err)
+	test.Assert(t, ok)
+	test.Assert(t, bizIntf.BizStatusCode() == testCode, bizIntf.BizStatusCode())
+	test.Assert(t, bizIntf.BizMessage() == testMsg, bizIntf.BizMessage())
+	test.DeepEqual(t, bizIntf.BizExtra(), testExtra)
+}
